Renames to BlobStream.
